ObsidianVault/scripts/requirements-enhancements.txt
# Long-term enhancement dependencies for PDF ingestion system
# Install with: pip install -r requirements-enhancements.txt

# OCR Support (Phase 1)
# Note: Requires Tesseract OCR to be installed separately
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Linux: sudo apt-get install tesseract-ocr
# macOS: brew install tesseract
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# Event-Driven Processing (Phase 1)
watchdog>=3.0.0

# AI/LLM Support
# OpenAI API for GPT models
openai>=1.0.0

# Anthropic API for Claude models
anthropic>=0.7.0

# Ollama for local LLM (optional, for privacy)
# Note: Requires Ollama to be installed separately
# https://ollama.ai/
ollama>=0.1.0

# Table Extraction
# Camelot for table extraction (requires ghostscript)
# Note: Install ghostscript separately
# Windows: https://www.ghostscript.com/download/gsdnld.html
# Linux: sudo apt-get install ghostscript
# macOS: brew install ghostscript
camelot-py[cv]>=0.11.0

# Tabula-py as alternative table extractor
# Note: Requires Java
tabula-py>=2.5.1

# Web API
# FastAPI for REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Additional web API dependencies
python-multipart>=0.0.6  # For file uploads
python-jose[cryptography]>=3.3.0  # For JWT authentication (optional)
passlib[bcrypt]>=1.7.4  # For password hashing (optional)

# Additional utilities
aiofiles>=23.2.1  # For async file operations
httpx>=0.25.0  # For async HTTP client
